MyAdventure - Added reset button and option hiding
Known issue: Buttons will be named "Not Available" when no options remain. If "Not Available" is clicked, the buttons will hide and disable. If the reset button is pressed they will return. I could not get the timing right to disable them when they reach the end but before the buttons are loaded.